Sparse LU Indexing

+++ RandomLU/LeftPermMat.m
@@ -2,7 +2,7 @@ function P=LeftPermMat(v)
 %LEFTPERMMAT Converts a left permutation matrix represented as vector 
 % back into a matrix
     I = eye(length(v));
-    P = sparse(I(v,:));
+    P = sparse(double(I(v,:)));
 end
